finally fixed this bs, wasted at least 3h on this.

+++ SDL Projekt1/material.h
@@ -4,9 +4,9 @@
 struct Material
 {
 	Vector3 col;
-	float reflection_index;
+	double reflection_index;
 
-	Material(Vector3 _col = Vector3(), float _reflection_index = 0) {
+	Material(Vector3 _col = Vector3(), double _reflection_index = 0) {
 		col = _col.get_normalized();
 		reflection_index = _reflection_index;
 	}
